backend/grocery_assistant/requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
drf-extra-fields==3.1.1
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drfpasswordless==1.5.7
flake8==3.9.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.6.0
isort==5.8.0
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.9
pillow==7.0.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytz==2021.1
requests==2.25.1
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
wcwidth==0.1.9
wheel==0.36.2
zipp==3.1.0
